Make a new space for input widgets
