+++ ryven/ironflow/node_interface/control.py
@@ -161,17 +161,3 @@ class NodeController(NodeInterfaceBase):
         self.draw()
 
 
-class SliderControl:
-    def __init__(self, gui: GUI, node: Node):
-        self.gui = gui
-        self.node = node
-        self.widget = widgets.FloatSlider(
-            value=self.node.val, min=0, max=10, continuous_update=False
-        )
-
-        self.widget.observe(self.widget_change, names="value")
-
-    def widget_change(self, change: dict) -> None:
-        self.node.set_state({"val": change["new"]}, 0)
-        self.node.update_event()
-        self.gui.flow_canvas.redraw()
